Add files via upload
The Instagram user analytics project is an effort to collect and analyze data about the activities of
Instagram users. The data collected will include information about the number of followers each
user has, the number of posts each user has made, the number of likes each post got, and the
information about photo tags. This information will be used to identify trends in how users use
Instagram and to suggest ways that businesses can maximize their impact on this social media
platform