Add files via upload
Operation Analytics is the analysis done for the complete end-to-end operations of a company.
With the help of this, the company then finds the areas in which it must improve upon. I will be
using MY SQL to analyze the data and to find different insights and patterns. I am going to find
insights about the number of jobs reviewed, the percentage shared by each language, weekly user
engagement, user growth, weekly user retention, and so on